backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.3.1
django-filter==25.2
djangorestframework==3.14.0
python-decouple==3.8
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.5
wheel==0.45.1
